fix: TiccleDetail change-button-style
